This is a much nicer way to write commit messages!
Use the full power of a text editor (VIMMMM).

And write detailed commitmessages that can be easily followed.

Use commands such as git reset HEAD~ to add to unfinished commits.

We can now use patterns such as
  commit -m "temp changes"
to save the current changes while moving on to another branch.

When checkking out another branch we can then do
  git log
to see if the last log was a temp change and git reset HEAD~ to continue
working on it!